project clean up
