+++ KarhooSDK/Api/Response/Availability.swift
@@ -22,21 +22,3 @@ public struct Availability: KarhooCodableModel {
         self.vehicles = (try? container.decode(VehicleAvailability.self, forKey: .vehicles)) ?? VehicleAvailability()
     }
 }
-
-public struct VehicleAvailability: KarhooCodableModel {
-
-    public let classes: [String]
-
-    public init(classes: [String] = []) {
-        self.classes = classes
-    }
-
-    enum CodingKeys: String, CodingKey {
-        case classes
-    }
-
-    public init(from decoder: Decoder) throws {
-        let container = try decoder.container(keyedBy: CodingKeys.self)
-        self.classes = (try? container.decode(Array.self, forKey: .classes)) ?? []
-    }
-}
